Refactoring. NOR -> ENG
Changed language for entity class from norwegian to english

+++ sjakkskjerm/src/main/java/no/ntnu/sjakkskjerm/message/MessageController.java
@@ -33,7 +33,7 @@ public class MessageController {
     @GetMapping
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @RequestMapping(path = "/specific")
-    public List<Message> getMessageTournament(@RequestParam int tournamentId) {
+    public List<Message> getMessageTournament(@RequestParam Long tournamentId) {
         System.out.println(tournamentId);
         return messageService.getMessageTournament(tournamentId);
     }
